Update project section images
